requirements.txt
asyncpg==0.29.0
hikari==2.0.0.dev122
hikari-arc==1.2.1
pre-commit==3.6.2
python-dotenv==1.0.1
ruff==0.2.2
SQLAlchemy==2.0.27